progress

+++ RecipeTraderDotNet.Core/Domain/Market/IMarket.cs
@@ -8,7 +8,7 @@ namespace RecipeTraderDotNet.Core.Domain.Market
         List<PublicRecipe> GetAllRecipes();
         string Publish(PrivateRecipe privateRecipe, decimal price);
         PrivateRecipe Purchase(int publicRecipeId, string requestUserId);
-        string Review(int publicRecipeId, string reviewerUserId, int rating, string comment);
+        string Review(int publicRecipeId, string reviewerUserId, double rating, string comment);
         string TakeDown(int publicRecipeId, string requestUserId);
         SystemInfo GetSystemInfo();
         string CreateUserMoneyAccount(string userId, decimal initBalance = 100);
